requirements.txt
pyyaml==6.0.1
python-dotenv==1.0.1
pandas==2.2.2
openai==2.24.0
Pillow==10.3.0
jsonschema==4.22.0
pdfplumber==0.11.4